+++ src/components/Layout/Content/BottomContent/BottomContent.js
@@ -13,17 +13,21 @@ import { setShowPublishModal } from '../../../../redux/slices/mainSlice'
 
 const BottomContent = () => {
   // set up useSelector to get value from store
+  const _map = useSelector((state) => state.mainSlice.map)
   const _searchResults = useSelector((state) => state.mainSlice.searchResults)
   const _clickResults = useSelector((state) => state.mainSlice.clickResults)
   const _searchLoading = useSelector((state) => state.mainSlice.searchLoading)
   const _showPublishModal = useSelector(
     (state) => state.mainSlice.showPublishModal
   )
+  const _showZoomNotice = useSelector((state) => state.mainSlice.showZoomNotice)
+
   // if you are setting redux state, call dispatch
   const dispatch = useDispatch()
 
   const ANALYZE_LINK = process.env.REACT_APP_ANALYZE_BTN_URL
   const SHOW_PUBLISH_BTN = process.env.REACT_APP_SHOW_PUBLISH_BTN
+  const MIN_ZOOM = process.env.REACT_APP_MIN_ZOOM_LEVEL
 
   function onAnalyzeClick() {
     window.open(ANALYZE_LINK, '_blank')
@@ -33,9 +37,21 @@ const BottomContent = () => {
     dispatch(setShowPublishModal(!_showPublishModal))
   }
 
+  function onZoomClick() {
+    _map.setZoom(MIN_ZOOM)
+  }
+
   return (
     <div className="BottomContent">
       <LeafMap></LeafMap>
+      {_showZoomNotice && (
+        <div className="ZoomNotice">
+          Images are not visible at this zoom level.
+          <a onClick={onZoomClick}>
+            <strong>Zoom in</strong>
+          </a>
+        </div>
+      )}
       <div className="actionButtons">
         {ANALYZE_LINK && (
           <button className="actionButton" onClick={() => onAnalyzeClick()}>