+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Make date/time field required
 - Date function to calculate the last 2 weeks
 - Make Publish button configurable in env variable
+- Remove BBOX button
 
 ### Added
 
@@ -32,6 +33,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Functionality for custom Logo and alt description
 - Env variable REACT_APP_SHOW_PUBLISH_BTN
 - FilmDrop - Element 84 logo
+- Minimum zoom level for search
+- Env variable REACT_APP_MIN_ZOOM_LEVEL
 
 ## v0.2.0 - 2023-Jan-13